+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+Unreleased
+----------
+
+-  feat: :doc:`/scripts/csvcut` adds an :code:`--ignore-unknown-columns` option to skip identifiers in :code:`-c/--columns` that do not match a column in the input.
+
 2.2.0 - December 15, 2025
 -------------------------